third_party/WebKit/LayoutTests/fast/css/font-face-add-same-family-later-expected.txt

Tests that adding a FontFace to existing family triggers font update.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS widthBeforeLoad != widthAfterLoad is true
PASS successfullyParsed is true

TEST COMPLETE

Test Text
